\name{HapMap}
\docType{data}
\alias{HapMap_geno}
\alias{HapMap_anno}
\title{Sample HapMap data set (r21)}
\description{
  This data set consists of the first 500 SNPs from chromosome 22 genotyped in the HapMap JPT+CHB population (N=90). 
}
\usage{HapMap_geno}
\format{\code{HapMap_geno} is a matrix with 500 rows and 90 columns. The genotypes are coded as follows: homozygote: 0 or 2, heterozygote: 1, missing data: 3.
\code{HapMap_anno} is a matrix with 500 rows and 10 columns containing some annotations of the SNPs. }
\source{www.hapmap.org}
\keyword{datasets}

